+++ db/migrate/20150806132514_add_publications_and_grants_to_preceptors.rb
@@ -1,0 +1,6 @@
+class AddPublicationsAndGrantsToPreceptors < ActiveRecord::Migration
+  def change
+    add_column :preceptors, :publications, :text
+    add_column :preceptors, :grants, :text
+  end
+end